Renames jumbo-dialog to environment-dialog
