+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/discussion/body/application/modify-thread-reply-dialog.tsx
@@ -6,7 +6,7 @@ import { AnyActionType } from "~/actions";
 import { bindActionCreators } from "redux";
 import CKEditor from "~/components/general/ckeditor";
 import Link from "~/components/general/link";
-import JumboDialog from "~/components/general/jumbo-dialog";
+import JumboDialog from "~/components/general/environment-dialog";
 import { modifyReplyFromCurrentThread, ModifyReplyFromCurrentThreadTriggerType } from "~/actions/main-function/discussion";
 import {StateType} from '~/reducers';
 
@@ -87,7 +87,7 @@ class ModifyThreadReply extends React.Component<ModifyThreadReplyProps, ModifyTh
     ]
     let footer = (closeDialog: ()=>any)=>{
       return (          
-         <div className="jumbo-dialog__button-container">
+         <div className="environment-dialog__button-container">
           <Link className="button button--warn button--standard-cancel" onClick={closeDialog} disabled={this.state.locked}>
           {this.props.i18n.text.get('plugin.discussion.createmessage.cancel')}
           </Link>
